requirements.txt
flask
gradio_client
gunicorn
Pillow
